+++ babel.config.json
@@ -1,6 +1,6 @@
 {
   "plugins": [
-    "@compiled/babel-plugin/experimental",
+    "@compiled/babel-plugin/unstable-static",
     ["@compiled/babel-plugin", { "nonce": "\"k0Mp1lEd\"", "importReact": false }]
   ],
   "presets": [
