+++ packages/react/src/unstable-static.tsx
