+++ test/test-imports.js
@@ -1,3 +1,4 @@
 require('@compiled/react');
 require('@compiled/react/runtime');
+require('@compiled/react/unstable-static');
 require('@compiled/babel-plugin');
