+++ tsconfig.paths.json
@@ -6,7 +6,7 @@
       "@compiled/css": ["packages/css/src/index.tsx"],
       "@compiled/jest": ["packages/jest/src/index.tsx"],
       "@compiled/react": ["packages/react/src/index.tsx"],
-      "@compiled/react/experimental": ["packages/react/src/experimental.tsx"],
+      "@compiled/react/unstable-static": ["packages/react/src/unstable-static.tsx"],
       "@compiled/react/jsx-dev-runtime": ["packages/react/src/jsx/jsx-dev-runtime.tsx"],
       "@compiled/react/jsx-runtime": ["packages/react/src/jsx/jsx-runtime.tsx"],
       "@compiled/react/runtime": ["packages/react/src/runtime.tsx"],
